requirements.txt
matplotlib==3.9.3
mpmath==1.3.0
numpy>=1.26.0
pandas==2.2.3
python-dotenv==1.0.1
comet-ml==3.47.6
tqdm==4.67.1
networkx==3.4.2
tensorflow
joblib
scikit-learn
torch
torchvision
seaborn
numpy 
matplotlib 
opencv-python
aedat
pyyaml==6.0.2


